update: 更新文件 ark_web_window_new_event_info_impl.cpp

+++ ohos_interface/ohos_glue/ohos_nweb/bridge/webcore/ark_web_window_new_event_info_impl.cpp
@@ -72,7 +72,7 @@ int32_t ArkWebWindowNewEventInfoImpl::GetHeight()
 
 int32_t ArkWebWindowNewEventInfoImpl::GetNavigationPolicy()
 {
-    return static_cast<size_t>(nweb_window_new_event_info_->GetNavigationPolicy());
+    return static_cast<int32_t>(nweb_window_new_event_info_->GetNavigationPolicy());
 }
 
 } // namespace OHOS::ArkWeb
